@@ -7,9 +7,13 @@ log = get_logger()
 
 # Does nothing currently
 from mailbag.derivative import Derivative
+
+
 class HtmlDerivative(Derivative):
-    derivative_name = 'html'
-    derivative_format = 'html'
+    derivative_name = "html"
+    derivative_format = "html"
+    derivative_agent = ""
+    derivative_agent_version = ""
 
     def __init__(self, email_account, **kwargs):
         log.debug("Setup account")
@@ -26,19 +30,43 @@ class HtmlDerivative(Derivative):
 
     def do_task_per_message(self, message):
 
-        out_dir = os.path.join(self.html_dir, message.Derivatives_Path)
-        filename = os.path.join(out_dir, str(message.Mailbag_Message_ID) + ".html")
+        errors = {}
+        errors["msg"] = []
+        errors["stack_trace"] = []
 
-        if message.HTML_Body is None and message.Text_Body is None:
-            log.warn("No HTML or plain text body for " + str(message.Mailbag_Message_ID) + ". No HTML derivative will be created.")
-        else:
-            log.debug("Writing html derivative to " + filename)
-            if not self.args.dry_run:
-                if not os.path.isdir(out_dir):
-                    os.makedirs(out_dir)
-                #Calling helper function to get formatted html
-                html_formatted, encoding = helper.htmlFormatting(message, self.args.css, headers=False)
-                with open(filename, "w", encoding=encoding) as f:
-                    f.write(html_formatted)
-                    f.close()
-                
+        try:
+
+            out_dir = os.path.join(self.html_dir, message.Derivatives_Path)
+            filename = os.path.join(out_dir, str(message.Mailbag_Message_ID) + ".html")
+
+            if message.HTML_Body is None and message.Text_Body is None:
+                desc = "No HTML or plain text body for " + str(message.Mailbag_Message_ID) + ", no HTML derivative created"
+                errors = helper.handle_error(errors, None, desc, "warn")
+            else:
+                log.debug("Writing html derivative to " + filename)
+                # Calling helper function to get formatted html
+                try:
+                    html_formatted, encoding = helper.htmlFormatting(message, self.args.css, headers=False)
+                except Exception as e:
+                    desc = "Error formatting HTML for HTML derivative"
+                    errors = helper.handle_error(errors, e, desc)
+
+                if not self.args.dry_run:
+                    try:
+                        if not os.path.isdir(out_dir):
+                            os.makedirs(out_dir)
+                        with open(filename, "w", encoding=encoding) as f:
+                            f.write(html_formatted)
+                            f.close()
+                    except Exception as e:
+                        desc = "Error writing HTML derivative"
+                        errors = helper.handle_error(errors, e, desc)
+
+        except Exception as e:
+            desc = "Error creating HTML derivative"
+            errors = helper.handle_error(errors, e, desc)
+
+        message.Error.extend(errors["msg"])
+        message.StackTrace.extend(errors["stack_trace"])
+
+        return message
