@@ -9,10 +9,10 @@ import mailbag.helper as helper
 # pst is not supported otherwise
 skip_registry = False
 try:
-    if distutils.spawn.find_executable('wkhtmltopdf.exe'):
-        wkhtmltopdf = 'wkhtmltopdf.exe'
-    elif distutils.spawn.find_executable('wkhtmltopdf'):
-        wkhtmltopdf = 'wkhtmltopdf'
+    if distutils.spawn.find_executable("wkhtmltopdf.exe"):
+        wkhtmltopdf = "wkhtmltopdf.exe"
+    elif distutils.spawn.find_executable("wkhtmltopdf"):
+        wkhtmltopdf = "wkhtmltopdf"
     else:
         skip_registry = True
         wkhtmltopdf = None
@@ -24,9 +24,16 @@ log = get_logger()
 if not skip_registry:
 
     class PDFDerivative(Derivative):
-        derivative_name = 'pdf'
-        derivative_format = 'pdf'
-        
+        derivative_name = "pdf"
+        derivative_format = "pdf"
+        derivative_agent = "wkhtmltopdf"
+        check_version = subprocess.Popen([wkhtmltopdf, "--version"], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        out, err = check_version.communicate()
+        if len(err) > 0:
+            log.error("Unable to access --version for " + derivative_agent + ": " + err.decode("utf-8"))
+            raise Exception("Unable to access --version for " + derivative_agent + ": " + err.decode("utf-8"))
+        derivative_agent_version = out.decode("utf-8").strip()
+
         def __init__(self, email_account, **kwargs):
             log.debug("Setup account")
             super()
@@ -42,38 +49,67 @@ if not skip_registry:
 
         def do_task_per_message(self, message):
 
-            out_dir = os.path.join(self.pdf_dir, message.Derivatives_Path)
-            filename = os.path.join(out_dir, str(message.Mailbag_Message_ID))
-            html_name = filename + ".html"
-            pdf_name = filename + ".pdf"
+            errors = {}
+            errors["msg"] = []
+            errors["stack_trace"] = []
 
-            if message.HTML_Body is None and message.Text_Body is None:
-                log.warn("No HTML or plain text body for " + str(message.Mailbag_Message_ID) + ". No PDF derivative will be created.")
-            else:
-                log.debug("Writing HTML to " + str(html_name) + " and converting to " + str(pdf_name))
-                html_formatted, encoding = helper.htmlFormatting(message, self.args.css)
+            try:
 
-                if not self.args.dry_run:
-                    if not os.path.isdir(out_dir):
-                        os.makedirs(out_dir)
+                out_dir = os.path.join(self.pdf_dir, message.Derivatives_Path)
+                filename = os.path.join(out_dir, str(message.Mailbag_Message_ID))
+                html_name = filename + ".html"
+                pdf_name = filename + ".pdf"
 
-                    with open(html_name, 'w', encoding=encoding) as write_html:
-                        write_html.write(html_formatted)
-                        write_html.close()
-                    p = subprocess.Popen([wkhtmltopdf, "--disable-javascript", html_name, pdf_name], stdout=subprocess.PIPE,
-                                         stderr=subprocess.PIPE)
+                if message.HTML_Body is None and message.Text_Body is None:
+                    desc = "No HTML or plain text body for " + str(message.Mailbag_Message_ID) + ", no PDF derivative created"
+                    errors = helper.handle_error(errors, None, desc, "warn")
+                else:
+                    log.debug("Writing HTML to " + str(html_name) + " and converting to " + str(pdf_name))
+                    # Calling helper function to get formatted html
+                    try:
+                        html_formatted, encoding = helper.htmlFormatting(message, self.args.css)
+                    except Exception as e:
+                        desc = "Error formatting HTML for PDF derivative"
+                        errors = helper.handle_error(errors, e, desc)
 
-                    stdout, stderr = p.communicate()
-                    if p.returncode == 0:
-                        log.debug("Successfully created " + str(message.Mailbag_Message_ID) + ".pdf")
-                    else:
-                        if stdout:
-                            log.warn("Output converting to " + str(message.Mailbag_Message_ID) + ".pdf: " + str(stdout))
-                        if stderr:
-                            log.error("Error converting to " + str(message.Mailbag_Message_ID) + ".pdf: " + str(stderr))
-                        # comment out for now since catches even very minor issues
-                        # raise TypeError(stderr)
-                    # delete the HTML file
-                    os.remove(html_name)
+                    if not self.args.dry_run:
+                        try:
+                            if not os.path.isdir(out_dir):
+                                os.makedirs(out_dir)
 
+                            with open(html_name, "w", encoding=encoding) as write_html:
+                                write_html.write(html_formatted)
+                                write_html.close()
+                            command = [
+                                wkhtmltopdf,
+                                "--disable-javascript",
+                                os.path.abspath(html_name),
+                                os.path.abspath(pdf_name),
+                            ]
+                            log.debug("Running " + " ".join(command))
+                            p = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+                            stdout, stderr = p.communicate()
+                            if p.returncode == 0:
+                                log.debug("Successfully created " + str(message.Mailbag_Message_ID) + ".pdf")
+                            else:
+                                if stdout:
+                                    log.debug("Output converting to " + str(message.Mailbag_Message_ID) + ".pdf: " + stdout.decode("utf-8"))
+                                if stderr:
+                                    desc = "Error converting to " + str(message.Mailbag_Message_ID) + ".pdf: " + stderr.decode("utf-8")
+                                    errors = helper.handle_error(errors, None, desc, "warn")
+                            # delete the HTML file
+                            if os.path.isfile(pdf_name):
+                                os.remove(html_name)
 
+                        except Exception as e:
+                            desc = "Error writing HTML and converting to PDF derivative"
+                            errors = helper.handle_error(errors, e, desc)
+
+            except Exception as e:
+                desc = "Error creating PDF derivative"
+                errors = helper.handle_error(errors, e, desc)
+
+            message.Error.extend(errors["msg"])
+            message.StackTrace.extend(errors["stack_trace"])
+
+            return message
