@@ -2,7 +2,6 @@ import datetime
 import json
 from os.path import join
 import mailbag.helper as helper
-import mailbox
 from structlog import get_logger
 from email import parser
 from mailbag.email_account import EmailAccount
@@ -10,13 +9,17 @@ from mailbag.models import Email, Attachment
 import email
 import glob, os
 from email import policy
+import platform
 
 log = get_logger()
 
 
 class EML(EmailAccount):
     """EML - This concrete class parses eml file format"""
-    format_name = 'eml'
+
+    format_name = "eml"
+    format_agent = email.__name__
+    format_agent_version = platform.python_version()
 
     def __init__(self, target_account, args, **kwargs):
         log.debug("Parsity parse")
@@ -37,11 +40,11 @@ class EML(EmailAccount):
         files = glob.glob(os.path.join(self.file, "**", "*.eml"), recursive=True)
 
         for filePath in files:
-            
+
             if self.iteration_only:
                 yield None
-                continue     
-            
+                continue
+
             originalFile = helper.relativePath(self.file, filePath)
 
             attachments = []
@@ -49,7 +52,7 @@ class EML(EmailAccount):
             errors["msg"] = []
             errors["stack_trace"] = []
             try:
-                with open(filePath, 'rb') as f:
+                with open(filePath, "rb") as f:
                     msg = email.message_from_binary_file(f, policy=policy.default)
 
                     try:
@@ -79,35 +82,31 @@ class EML(EmailAccount):
                         errors = helper.handle_error(errors, e, desc)
 
                     message = Email(
-                            Error=errors["msg"],
-                            Message_ID=msg["message-id"].strip(),
-                            Original_File=originalFile,
-                            Message_Path=messagePath,
-                            Derivatives_Path=derivativesPath,
-                            Date=msg["date"],
-                            From=msg["from"],
-                            To=msg["to"],
-                            Subject=msg["subject"],
-                            Content_Type=msg.get_content_type(),
-                            Headers=msg,
-                            HTML_Body=bodies["html_body"],
-                            HTML_Encoding=bodies["html_encoding"],
-                            Text_Body=bodies["text_body"],
-                            Text_Encoding=bodies["text_encoding"],
-                            Message=msg,
-                            Attachments=attachments,
-                            StackTrace=errors["stack_trace"]
-                        )
+                        Error=errors["msg"],
+                        Message_ID=msg["message-id"].strip(),
+                        Original_File=originalFile,
+                        Message_Path=messagePath,
+                        Derivatives_Path=derivativesPath,
+                        Date=msg["date"],
+                        From=msg["from"],
+                        To=msg["to"],
+                        Subject=msg["subject"],
+                        Content_Type=msg.get_content_type(),
+                        Headers=msg,
+                        HTML_Body=bodies["html_body"],
+                        HTML_Encoding=bodies["html_encoding"],
+                        Text_Body=bodies["text_body"],
+                        Text_Encoding=bodies["text_encoding"],
+                        Message=msg,
+                        Attachments=attachments,
+                        StackTrace=errors["stack_trace"],
+                    )
 
             except (email.errors.MessageParseError, Exception) as e:
-                desc = 'Error parsing message'
+                desc = "Error parsing message"
                 errors = helper.handle_error(errors, e, desc)
-                message = Email(
-                    Error=errors["msg"],
-                    StackTrace=errors["stack_trace"]
-                )
-
+                message = Email(Error=errors["msg"], StackTrace=errors["stack_trace"])
 
             # Move EML to new mailbag directory structure
             yield message
-            new_path = helper.moveWithDirectoryStructure(self.dry_run,self.file,self.mailbag_name,self.format_name,filePath)
+            new_path = helper.moveWithDirectoryStructure(self.dry_run, self.file, self.mailbag_name, self.format_name, filePath)
